Reworking data fetching. Removing Redux
