+++ utils/devCatalog.js
@@ -1,0 +1,44 @@
+import JSONBig from 'json-bigint';
+import { Client, Environment } from 'square';
+
+export const devCatalog = async () => {
+	const client = new Client({
+		environment: Environment.Production,
+		accessToken: process.env.SQUARE_ACCESS_TOKEN,
+	});
+
+	const newImageRequest = async (items) => {
+		const catalog = client.catalogApi;
+
+		let newItemsWithPictures = [];
+
+		for (let i = 0; i < items.length; i++) {
+			const response = await catalog.retrieveCatalogObject(items[i].imageId);
+			items[i].imageLink = response.result.object.imageData.url;
+			newItemsWithPictures.push(items[i]);
+		}
+		return newItemsWithPictures;
+	};
+
+	let items = [];
+	let filteredItems = [];
+
+	const catalog = client.catalogApi;
+	const response = await catalog.listCatalog('', 'ITEM');
+	items = JSONBig.parse(JSONBig.stringify(response.result.objects));
+
+	//Then the items are filtered so that only ones that have photo's are returned
+	if (items) {
+		const dataItems = items;
+		for (let i = 0; i < dataItems.length; i++) {
+			if (dataItems[i].imageId) {
+				filteredItems.push(dataItems[i]);
+			}
+		}
+
+		//Finally, before returning the list of Items it grabs the URL for the photo's for each item -- This takes a while!
+		const itemsWithPictures = await newImageRequest(filteredItems);
+
+		return itemsWithPictures;
+	}
+};
